Module wasmer_wasix_types::wasi::bindings

Structs

AddrUnspec

AddrUnspecPort

CidrUnspec

Dirent

A directory entry.

ErrnoSignal

Represents an errno and a signal

EventFdReadwrite

The contents of an event for the eventtype::fd_read and eventtype::fd_write variants

Eventrwflags

The state of the file descriptor subscribed to with eventtype::fd_read or eventtype::fd_write.

Fdflags

File descriptor flags.

Fdstat

File descriptor attributes.

Filestat

Fstflags

Which file time attributes to adjust. TODO: wit appears to not have support for flags repr (@witx repr u16)

HttpHandles

HttpStatus

JoinFlags

join flags.

Lookup

Flags determining the method of how paths are resolved. TODO: wit appears to not have support for flags repr (@witx repr u32)

Oflags

Open flags used by path_open. TODO: wit appears to not have support for flags repr (@witx repr u16)

OptionFd

OptionPid

OptionTimestamp

PipeHandles

Prestat

PrestatU

PrestatUDir

ProcessHandles

Rights

File descriptor rights, determining which actions may be performed.

Snapshot0Dirent

A directory entry, snapshot0 version.

Snapshot0Filestat

Snapshot0SubscriptionClock

The contents of a subscription when type is eventtype::clock.

Subclockflags

Flags determining how to interpret the timestamp provided in subscription-clock::timeout.

SubscriptionClock

The contents of a subscription when type is eventtype::clock.

SubscriptionFsReadwrite

An event that occurred. The contents of an event. An event that occurred. The contents of a subscription, snapshot0 version. The contents of a subscription. The contents of a subscription when the variant is eventtype::fd_read or eventtype::fd_write.

ThreadStateFlags

thread state flags

Tty

Enums

Addressfamily

Advice

File or memory access pattern advisory information.

Bool

Clockid

Identifiers for clocks.

Errno

Error codes returned by functions. Not all of these error codes are returned by the functions provided by this API; some are used in higher-level library layers, and others are provided merely for alignment with POSIX.

Eventtype

Type of a subscription to an event or its occurrence.

Filetype

The type of a file descriptor or file.

JoinStatusType

What has happened with the proccess when we joined on it

OptionTag

__wasi_option_t

Preopentype

Identifiers for preopened capabilities.

Signal

Snapshot0Clockid

Identifiers for clocks, snapshot0 version.

Snapshot0Whence

SockProto

Sockoption

Sockstatus

Socktype

StdioMode

Streamsecurity

Timeout

Whence

Constants

Access__num_enum_0__ ğŸ”’

Addrinuse__num_enum_0__ ğŸ”’

Addrnotavail__num_enum_0__ ğŸ”’

Afnosupport__num_enum_0__ ğŸ”’

Again__num_enum_0__ ğŸ”’

Ah__num_enum_0__ ğŸ”’

Already__num_enum_0__ ğŸ”’

Badf__num_enum_0__ ğŸ”’

Badmsg__num_enum_0__ ğŸ”’

Beetph__num_enum_0__ ğŸ”’

Busy__num_enum_0__ ğŸ”’

Canceled__num_enum_0__ ğŸ”’

Child__num_enum_0__ ğŸ”’

Comp__num_enum_0__ ğŸ”’

Connaborted__num_enum_0__ ğŸ”’

Connrefused__num_enum_0__ ğŸ”’

Connreset__num_enum_0__ ğŸ”’

Dccp__num_enum_0__ ğŸ”’

Deadlk__num_enum_0__ ğŸ”’

Destaddrreq__num_enum_0__ ğŸ”’

Dom__num_enum_0__ ğŸ”’

Dquot__num_enum_0__ ğŸ”’

Dstopts__num_enum_0__ ğŸ”’

Egp__num_enum_0__ ğŸ”’

Encap__num_enum_0__ ğŸ”’

Esp__num_enum_0__ ğŸ”’

Ethernet__num_enum_0__ ğŸ”’

Exist__num_enum_0__ ğŸ”’

Fault__num_enum_0__ ğŸ”’

Fbig__num_enum_0__ ğŸ”’

Fragment__num_enum_0__ ğŸ”’

Gre__num_enum_0__ ğŸ”’

Hostunreach__num_enum_0__ ğŸ”’

Icmp__num_enum_0__ ğŸ”’

Icmpv6__num_enum_0__ ğŸ”’

Idp__num_enum_0__ ğŸ”’

Idrm__num_enum_0__ ğŸ”’

Igmp__num_enum_0__ ğŸ”’

Ilseq__num_enum_0__ ğŸ”’

Inprogress__num_enum_0__ ğŸ”’

Intr__num_enum_0__ ğŸ”’

Inval__num_enum_0__ ğŸ”’

Io__num_enum_0__ ğŸ”’

Ip__num_enum_0__ ğŸ”’

Ipip__num_enum_0__ ğŸ”’

Ipv6__num_enum_0__ ğŸ”’

Isconn__num_enum_0__ ğŸ”’

Isdir__num_enum_0__ ğŸ”’

Loop__num_enum_0__ ğŸ”’

Max__num_enum_0__ ğŸ”’

Memviolation__num_enum_0__ ğŸ”’

Mfile__num_enum_0__ ğŸ”’

Mh__num_enum_0__ ğŸ”’

Mlink__num_enum_0__ ğŸ”’

Monotonic__num_enum_0__ ğŸ”’

Monotonic__num_enum_0__ ğŸ”’

Mpls__num_enum_0__ ğŸ”’

Mptcp__num_enum_0__ ğŸ”’

Msgsize__num_enum_0__ ğŸ”’

Mtp__num_enum_0__ ğŸ”’

Multihop__num_enum_0__ ğŸ”’

Nametoolong__num_enum_0__ ğŸ”’

Netdown__num_enum_0__ ğŸ”’

Netreset__num_enum_0__ ğŸ”’

Netunreach__num_enum_0__ ğŸ”’

Nfile__num_enum_0__ ğŸ”’

Nobufs__num_enum_0__ ğŸ”’

Nodev__num_enum_0__ ğŸ”’

Noent__num_enum_0__ ğŸ”’

Noexec__num_enum_0__ ğŸ”’

Nolck__num_enum_0__ ğŸ”’

Nolink__num_enum_0__ ğŸ”’

Nomem__num_enum_0__ ğŸ”’

Nomsg__num_enum_0__ ğŸ”’

None__num_enum_0__ ğŸ”’

Noprotoopt__num_enum_0__ ğŸ”’

Nospc__num_enum_0__ ğŸ”’

Nosys__num_enum_0__ ğŸ”’

Notcapable__num_enum_0__ ğŸ”’

Notconn__num_enum_0__ ğŸ”’

Notdir__num_enum_0__ ğŸ”’

Notempty__num_enum_0__ ğŸ”’

Notrecoverable__num_enum_0__ ğŸ”’

Notsock__num_enum_0__ ğŸ”’

Notsup__num_enum_0__ ğŸ”’

Notty__num_enum_0__ ğŸ”’

Nxio__num_enum_0__ ğŸ”’

Overflow__num_enum_0__ ğŸ”’

Ownerdead__num_enum_0__ ğŸ”’

Perm__num_enum_0__ ğŸ”’

Pim__num_enum_0__ ğŸ”’

Pipe__num_enum_0__ ğŸ”’

ProcessCputimeId__num_enum_0__ ğŸ”’

ProcessCputimeId__num_enum_0__ ğŸ”’

ProtoEighteen__num_enum_0__ ğŸ”’

ProtoEighty__num_enum_0__ ğŸ”’

ProtoEightyeight__num_enum_0__ ğŸ”’

ProtoEightyfive__num_enum_0__ ğŸ”’

ProtoEightyfour__num_enum_0__ ğŸ”’

ProtoEightynine__num_enum_0__ ğŸ”’

ProtoEightyone__num_enum_0__ ğŸ”’

ProtoEightyseven__num_enum_0__ ğŸ”’

ProtoEightysix__num_enum_0__ ğŸ”’

ProtoEightythree__num_enum_0__ ğŸ”’

ProtoEightytwo__num_enum_0__ ğŸ”’

ProtoEleven__num_enum_0__ ğŸ”’

ProtoFifteen__num_enum_0__ ğŸ”’

ProtoFiftyfive__num_enum_0__ ğŸ”’

ProtoFiftyfour__num_enum_0__ ğŸ”’

ProtoFiftyseven__num_enum_0__ ğŸ”’

ProtoFiftysix__num_enum_0__ ğŸ”’

ProtoFiftythree__num_enum_0__ ğŸ”’

ProtoFiftytwo__num_enum_0__ ğŸ”’

ProtoFive__num_enum_0__ ğŸ”’

ProtoFourteen__num_enum_0__ ğŸ”’

ProtoFourty__num_enum_0__ ğŸ”’

ProtoFourtyeight__num_enum_0__ ğŸ”’

ProtoFourtyfive__num_enum_0__ ğŸ”’

ProtoFourtynine__num_enum_0__ ğŸ”’

ProtoFourtytwo__num_enum_0__ ğŸ”’

ProtoNine__num_enum_0__ ğŸ”’

ProtoNineetyseven__num_enum_0__ ğŸ”’

ProtoNineteen__num_enum_0__ ğŸ”’

ProtoNinety__num_enum_0__ ğŸ”’

ProtoNinetyfive__num_enum_0__ ğŸ”’

ProtoNinetynine__num_enum_0__ ğŸ”’

ProtoNinetyone__num_enum_0__ ğŸ”’

ProtoNinetysix__num_enum_0__ ğŸ”’

ProtoNinetythree__num_enum_0__ ğŸ”’

ProtoOnehundred__num_enum_0__ ğŸ”’

ProtoOnehundredandeighteen__num_enum_0__ ğŸ”’

ProtoOnehundredandeighty__num_enum_0__ ğŸ”’

ProtoOnehundredandeightyeight__num_enum_0__ ğŸ”’

ProtoOnehundredandeightyfive__num_enum_0__ ğŸ”’

ProtoOnehundredandeightyfour__num_enum_0__ ğŸ”’

ProtoOnehundredandeightynine__num_enum_0__ ğŸ”’

ProtoOnehundredandeightyone__num_enum_0__ ğŸ”’

ProtoOnehundredandeightyseven__num_enum_0__ ğŸ”’

ProtoOnehundredandeightysix__num_enum_0__ ğŸ”’

ProtoOnehundredandeightythree__num_enum_0__ ğŸ”’

ProtoOnehundredandeightytwo__num_enum_0__ ğŸ”’

ProtoOnehundredandeleven__num_enum_0__ ğŸ”’

ProtoOnehundredandfifteen__num_enum_0__ ğŸ”’

ProtoOnehundredandfifty__num_enum_0__ ğŸ”’

ProtoOnehundredandfiftyeight__num_enum_0__ ğŸ”’

ProtoOnehundredandfiftyfive__num_enum_0__ ğŸ”’

ProtoOnehundredandfiftyfour__num_enum_0__ ğŸ”’

ProtoOnehundredandfiftynine__num_enum_0__ ğŸ”’

ProtoOnehundredandfiftyone__num_enum_0__ ğŸ”’

ProtoOnehundredandfiftyseven__num_enum_0__ ğŸ”’

ProtoOnehundredandfiftysix__num_enum_0__ ğŸ”’

ProtoOnehundredandfiftythree__num_enum_0__ ğŸ”’

ProtoOnehundredandfiftytwo__num_enum_0__ ğŸ”’

ProtoOnehundredandfive__num_enum_0__ ğŸ”’

ProtoOnehundredandfour__num_enum_0__ ğŸ”’

ProtoOnehundredandfourteen__num_enum_0__ ğŸ”’

ProtoOnehundredandfourty__num_enum_0__ ğŸ”’

ProtoOnehundredandfourtyeight__num_enum_0__ ğŸ”’

ProtoOnehundredandfourtyfive__num_enum_0__ ğŸ”’

ProtoOnehundredandfourtyfour__num_enum_0__ ğŸ”’

ProtoOnehundredandfourtynine__num_enum_0__ ğŸ”’

ProtoOnehundredandfourtyone__num_enum_0__ ğŸ”’

ProtoOnehundredandfourtyseven__num_enum_0__ ğŸ”’

ProtoOnehundredandfourtysix__num_enum_0__ ğŸ”’

ProtoOnehundredandfourtytwo__num_enum_0__ ğŸ”’

ProtoOnehundredandnine__num_enum_0__ ğŸ”’

ProtoOnehundredandnineteen__num_enum_0__ ğŸ”’

ProtoOnehundredandninety__num_enum_0__ ğŸ”’

ProtoOnehundredandninetyeight__num_enum_0__ ğŸ”’

ProtoOnehundredandninetyfive__num_enum_0__ ğŸ”’

ProtoOnehundredandninetyfour__num_enum_0__ ğŸ”’

ProtoOnehundredandninetynine__num_enum_0__ ğŸ”’

ProtoOnehundredandninetyone__num_enum_0__ ğŸ”’

ProtoOnehundredandninetyseven__num_enum_0__ ğŸ”’

ProtoOnehundredandninetysix__num_enum_0__ ğŸ”’

ProtoOnehundredandninetythree__num_enum_0__ ğŸ”’

ProtoOnehundredandninetytwo__num_enum_0__ ğŸ”’

ProtoOnehundredandone__num_enum_0__ ğŸ”’

ProtoOnehundredandseven__num_enum_0__ ğŸ”’

ProtoOnehundredandseventeen__num_enum_0__ ğŸ”’

ProtoOnehundredandseventy__num_enum_0__ ğŸ”’

ProtoOnehundredandseventyeight__num_enum_0__ ğŸ”’

ProtoOnehundredandseventyfive__num_enum_0__ ğŸ”’

ProtoOnehundredandseventyfour__num_enum_0__ ğŸ”’

ProtoOnehundredandseventynine__num_enum_0__ ğŸ”’

ProtoOnehundredandseventyone__num_enum_0__ ğŸ”’

ProtoOnehundredandseventyseven__num_enum_0__ ğŸ”’

ProtoOnehundredandseventysix__num_enum_0__ ğŸ”’

ProtoOnehundredandseventythree__num_enum_0__ ğŸ”’

ProtoOnehundredandseventytwo__num_enum_0__ ğŸ”’

ProtoOnehundredandsix__num_enum_0__ ğŸ”’

ProtoOnehundredandsixteen__num_enum_0__ ğŸ”’

ProtoOnehundredandsixty__num_enum_0__ ğŸ”’

ProtoOnehundredandsixtyeight__num_enum_0__ ğŸ”’

ProtoOnehundredandsixtyfive__num_enum_0__ ğŸ”’

ProtoOnehundredandsixtyfour__num_enum_0__ ğŸ”’

ProtoOnehundredandsixtynine__num_enum_0__ ğŸ”’

ProtoOnehundredandsixtyone__num_enum_0__ ğŸ”’

ProtoOnehundredandsixtyseven__num_enum_0__ ğŸ”’

ProtoOnehundredandsixtysix__num_enum_0__ ğŸ”’

ProtoOnehundredandsixtythree__num_enum_0__ ğŸ”’

ProtoOnehundredandsixtytwo__num_enum_0__ ğŸ”’

ProtoOnehundredandten__num_enum_0__ ğŸ”’

ProtoOnehundredandthirteen__num_enum_0__ ğŸ”’

ProtoOnehundredandthirty__num_enum_0__ ğŸ”’

ProtoOnehundredandthirtyeight__num_enum_0__ ğŸ”’

ProtoOnehundredandthirtyfour__num_enum_0__ ğŸ”’

ProtoOnehundredandthirtynine__num_enum_0__ ğŸ”’

ProtoOnehundredandthirtyone__num_enum_0__ ğŸ”’

ProtoOnehundredandthirtythree__num_enum_0__ ğŸ”’

ProtoOnehundredandtwelve__num_enum_0__ ğŸ”’

ProtoOnehundredandtwenty__num_enum_0__ ğŸ”’

ProtoOnehundredandtwentyeight__num_enum_0__ ğŸ”’

ProtoOnehundredandtwentyfive__num_enum_0__ ğŸ”’

ProtoOnehundredandtwentyfour__num_enum_0__ ğŸ”’

ProtoOnehundredandtwentynine__num_enum_0__ ğŸ”’

ProtoOnehundredandtwentyone__num_enum_0__ ğŸ”’

ProtoOnehundredandtwentyseven__num_enum_0__ ğŸ”’

ProtoOnehundredandtwentysix__num_enum_0__ ğŸ”’

ProtoOnehundredandtwentythree__num_enum_0__ ğŸ”’

ProtoOnehundredandtwentytwo__num_enum_0__ ğŸ”’

ProtoOnehundredandtwo__num_enum_0__ ğŸ”’

ProtoRaw__num_enum_0__ ğŸ”’

ProtoSeven__num_enum_0__ ğŸ”’

ProtoSeventy__num_enum_0__ ğŸ”’

ProtoSeventyeight__num_enum_0__ ğŸ”’

ProtoSeventyfive__num_enum_0__ ğŸ”’

ProtoSeventyfour__num_enum_0__ ğŸ”’

ProtoSeventynine__num_enum_0__ ğŸ”’

ProtoSeventyone__num_enum_0__ ğŸ”’

ProtoSeventyseven__num_enum_0__ ğŸ”’

ProtoSeventysix__num_enum_0__ ğŸ”’

ProtoSeventythree__num_enum_0__ ğŸ”’

ProtoSeventytwo__num_enum_0__ ğŸ”’

ProtoSixteen__num_enum_0__ ğŸ”’

ProtoSixtyeight__num_enum_0__ ğŸ”’

ProtoSixtyfive__num_enum_0__ ğŸ”’

ProtoSixtyfour__num_enum_0__ ğŸ”’

ProtoSixtynine__num_enum_0__ ğŸ”’

ProtoSixtyone__num_enum_0__ ğŸ”’

ProtoSixtyseven__num_enum_0__ ğŸ”’

ProtoSixtysix__num_enum_0__ ğŸ”’

ProtoSixtythree__num_enum_0__ ğŸ”’

ProtoSixtytwo__num_enum_0__ ğŸ”’

ProtoTen__num_enum_0__ ğŸ”’

ProtoThirteen__num_enum_0__ ğŸ”’

ProtoThirty__num_enum_0__ ğŸ”’

ProtoThirtyeight__num_enum_0__ ğŸ”’

ProtoThirtyfive__num_enum_0__ ğŸ”’

ProtoThirtyfour__num_enum_0__ ğŸ”’

ProtoThirtynine__num_enum_0__ ğŸ”’

ProtoThirtyone__num_enum_0__ ğŸ”’

ProtoThirtyseven__num_enum_0__ ğŸ”’

ProtoThirtysix__num_enum_0__ ğŸ”’

ProtoThirtytwo__num_enum_0__ ğŸ”’

ProtoThree__num_enum_0__ ğŸ”’

ProtoTp__num_enum_0__ ğŸ”’

ProtoTwenty__num_enum_0__ ğŸ”’

ProtoTwentyeight__num_enum_0__ ğŸ”’

ProtoTwentyfive__num_enum_0__ ğŸ”’

ProtoTwentyfour__num_enum_0__ ğŸ”’

ProtoTwentyone__num_enum_0__ ğŸ”’

ProtoTwentyseven__num_enum_0__ ğŸ”’

ProtoTwentysix__num_enum_0__ ğŸ”’

ProtoTwentythree__num_enum_0__ ğŸ”’

ProtoTwohundred__num_enum_0__ ğŸ”’

ProtoTwohundredandeight__num_enum_0__ ğŸ”’

ProtoTwohundredandeighteen__num_enum_0__ ğŸ”’

ProtoTwohundredandeleven__num_enum_0__ ğŸ”’

ProtoTwohundredandfifteen__num_enum_0__ ğŸ”’

ProtoTwohundredandfifty__num_enum_0__ ğŸ”’

ProtoTwohundredandfiftyeight__num_enum_0__ ğŸ”’

ProtoTwohundredandfiftyfour__num_enum_0__ ğŸ”’

ProtoTwohundredandfiftynine__num_enum_0__ ğŸ”’

ProtoTwohundredandfiftyone__num_enum_0__ ğŸ”’

ProtoTwohundredandfiftyseven__num_enum_0__ ğŸ”’

ProtoTwohundredandfiftysix__num_enum_0__ ğŸ”’

ProtoTwohundredandfiftythree__num_enum_0__ ğŸ”’

ProtoTwohundredandfiftytwo__num_enum_0__ ğŸ”’

ProtoTwohundredandfive__num_enum_0__ ğŸ”’

ProtoTwohundredandfour__num_enum_0__ ğŸ”’

ProtoTwohundredandfourteen__num_enum_0__ ğŸ”’

ProtoTwohundredandfourty__num_enum_0__ ğŸ”’

ProtoTwohundredandfourtyeight__num_enum_0__ ğŸ”’

ProtoTwohundredandfourtyfive__num_enum_0__ ğŸ”’

ProtoTwohundredandfourtyfour__num_enum_0__ ğŸ”’

ProtoTwohundredandfourtynine__num_enum_0__ ğŸ”’

ProtoTwohundredandfourtyone__num_enum_0__ ğŸ”’

ProtoTwohundredandfourtyseven__num_enum_0__ ğŸ”’

ProtoTwohundredandfourtysix__num_enum_0__ ğŸ”’

ProtoTwohundredandfourtythree__num_enum_0__ ğŸ”’

ProtoTwohundredandfourtytwo__num_enum_0__ ğŸ”’

ProtoTwohundredandnine__num_enum_0__ ğŸ”’

ProtoTwohundredandnineteen__num_enum_0__ ğŸ”’

ProtoTwohundredandone__num_enum_0__ ğŸ”’

ProtoTwohundredandseven__num_enum_0__ ğŸ”’

ProtoTwohundredandseventeen__num_enum_0__ ğŸ”’

ProtoTwohundredandsix__num_enum_0__ ğŸ”’

ProtoTwohundredandsixteen__num_enum_0__ ğŸ”’

ProtoTwohundredandsixty__num_enum_0__ ğŸ”’

ProtoTwohundredandsixtyone__num_enum_0__ ğŸ”’

ProtoTwohundredandten__num_enum_0__ ğŸ”’

ProtoTwohundredandthirteen__num_enum_0__ ğŸ”’

ProtoTwohundredandthirty__num_enum_0__ ğŸ”’

ProtoTwohundredandthirtyeight__num_enum_0__ ğŸ”’

ProtoTwohundredandthirtyfive__num_enum_0__ ğŸ”’

ProtoTwohundredandthirtyfour__num_enum_0__ ğŸ”’

ProtoTwohundredandthirtynine__num_enum_0__ ğŸ”’

ProtoTwohundredandthirtyone__num_enum_0__ ğŸ”’

ProtoTwohundredandthirtyseven__num_enum_0__ ğŸ”’

ProtoTwohundredandthirtysix__num_enum_0__ ğŸ”’

ProtoTwohundredandthirtythree__num_enum_0__ ğŸ”’

ProtoTwohundredandthirtytwo__num_enum_0__ ğŸ”’

ProtoTwohundredandthree__num_enum_0__ ğŸ”’

ProtoTwohundredandtwelve__num_enum_0__ ğŸ”’

ProtoTwohundredandtwenty__num_enum_0__ ğŸ”’

ProtoTwohundredandtwentyeight__num_enum_0__ ğŸ”’

ProtoTwohundredandtwentyfive__num_enum_0__ ğŸ”’

ProtoTwohundredandtwentyfour__num_enum_0__ ğŸ”’

ProtoTwohundredandtwentynine__num_enum_0__ ğŸ”’

ProtoTwohundredandtwentyone__num_enum_0__ ğŸ”’

ProtoTwohundredandtwentyseven__num_enum_0__ ğŸ”’

ProtoTwohundredandtwentysix__num_enum_0__ ğŸ”’

ProtoTwohundredandtwentythree__num_enum_0__ ğŸ”’

ProtoTwohundredandtwentytwo__num_enum_0__ ğŸ”’

ProtoTwohundredandtwo__num_enum_0__ ğŸ”’

Proto__num_enum_0__ ğŸ”’

Protonosupport__num_enum_0__ ğŸ”’

Prototype__num_enum_0__ ğŸ”’

Pup__num_enum_0__ ğŸ”’

Range__num_enum_0__ ğŸ”’

Realtime__num_enum_0__ ğŸ”’

Realtime__num_enum_0__ ğŸ”’

Rofs__num_enum_0__ ğŸ”’

Routing__num_enum_0__ ğŸ”’

Rsvp__num_enum_0__ ğŸ”’

Sctp__num_enum_0__ ğŸ”’

Shutdown__num_enum_0__ ğŸ”’

Sigabrt__num_enum_0__ ğŸ”’

Sigalrm__num_enum_0__ ğŸ”’

Sigbus__num_enum_0__ ğŸ”’

Sigchld__num_enum_0__ ğŸ”’

Sigcont__num_enum_0__ ğŸ”’

Sigfpe__num_enum_0__ ğŸ”’

Sighup__num_enum_0__ ğŸ”’

Sigill__num_enum_0__ ğŸ”’

Sigint__num_enum_0__ ğŸ”’

Sigkill__num_enum_0__ ğŸ”’

Signone__num_enum_0__ ğŸ”’

Sigpipe__num_enum_0__ ğŸ”’

Sigpoll__num_enum_0__ ğŸ”’

Sigprof__num_enum_0__ ğŸ”’

Sigpwr__num_enum_0__ ğŸ”’

Sigquit__num_enum_0__ ğŸ”’

Sigsegv__num_enum_0__ ğŸ”’

Sigstkflt__num_enum_0__ ğŸ”’

Sigstop__num_enum_0__ ğŸ”’

Sigsys__num_enum_0__ ğŸ”’

Sigterm__num_enum_0__ ğŸ”’

Sigtrap__num_enum_0__ ğŸ”’

Sigtstp__num_enum_0__ ğŸ”’

Sigttin__num_enum_0__ ğŸ”’

Sigttou__num_enum_0__ ğŸ”’

Sigurg__num_enum_0__ ğŸ”’

Sigusr1__num_enum_0__ ğŸ”’

Sigusr2__num_enum_0__ ğŸ”’

Sigvtalrm__num_enum_0__ ğŸ”’

Sigwinch__num_enum_0__ ğŸ”’

Sigxcpu__num_enum_0__ ğŸ”’

Sigxfsz__num_enum_0__ ğŸ”’

Spipe__num_enum_0__ ğŸ”’

Srch__num_enum_0__ ğŸ”’

Stale__num_enum_0__ ğŸ”’

Success__num_enum_0__ ğŸ”’

Tcp__num_enum_0__ ğŸ”’

ThreadCputimeId__num_enum_0__ ğŸ”’

ThreadCputimeId__num_enum_0__ ğŸ”’

Timedout__num_enum_0__ ğŸ”’

Toobig__num_enum_0__ ğŸ”’

Txtbsy__num_enum_0__ ğŸ”’

Udp__num_enum_0__ ğŸ”’

Udplite__num_enum_0__ ğŸ”’

Unknown__num_enum_0__ ğŸ”’

Unknown__num_enum_0__ ğŸ”’

Unknown__num_enum_0__ ğŸ”’

Xdev__num_enum_0__ ğŸ”’

Statics

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

Traits

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

Type Definitions

Count

Device

Identifier for a device containing a file system. Can be used in combination with inode to uniquely identify a file or directory in the filesystem.

Dircookie

A reference to the offset of a directory entry.

Dirnamlen

The type for the dirent::d-namlen field of dirent struct.

EventFdFlags

Fd

A file descriptor handle.

FileDelta

Filesize

Non-negative file size or length of a region within a file.

Inode

File serial number that is unique within its file system.

Linkcount

LookupFlags

Pid

RiFlags

RoFlags

SdFlags

SiFlags

Size

Type names used by low-level WASI interfaces. An array size.

Snapshot0Linkcount

Tid

Timestamp

Timestamp in nanoseconds.

Userdata

User-provided value that may be attached to objects that is retained when extracted from the implementation.