Struct wasmer_wasix_types::wasi::Rights

pub struct Rights {
    bits: u64,
}

File descriptor rights, determining which actions may be performed.

Fields
bits: u64
Implementations

impl Rights

pub const FD_DATASYNC: Self = _

The right to invoke fd_datasync.

If rights::path_open is set, includes the right to invoke path_open with fdflags::dsync.

pub const FD_READ: Self = _

The right to invoke fd_read and sock_recv.

If rights::fd_seek is set, includes the right to invoke fd_pread.

pub const FD_SEEK: Self = _

The right to invoke fd_seek. This flag implies rights::fd_tell.

pub const FD_FDSTAT_SET_FLAGS: Self = _

The right to invoke fd_fdstat_set_flags.

pub const FD_SYNC: Self = _

The right to invoke fd_sync.

If rights::path_open is set, includes the right to invoke path_open with fdflags::rsync and fdflags::dsync.

pub const FD_TELL: Self = _

The right to invoke fd_seek in such a way that the file offset remains unaltered (i.e., whence::cur with offset zero), or to invoke fd_tell.

pub const FD_WRITE: Self = _

The right to invoke fd_write and sock_send. If rights::fd_seek is set, includes the right to invoke fd_pwrite.

pub const FD_ADVISE: Self = _

The right to invoke fd_advise.

pub const FD_ALLOCATE: Self = _

The right to invoke fd_allocate.

pub const PATH_CREATE_DIRECTORY: Self = _

The right to invoke path_create_directory.

pub const PATH_CREATE_FILE: Self = _

If rights::path_open is set, the right to invoke path_open with oflags::creat.

pub const PATH_LINK_SOURCE: Self = _

The right to invoke path_link with the file descriptor as the source directory.

pub const PATH_LINK_TARGET: Self = _

The right to invoke path_link with the file descriptor as the target directory.

pub const PATH_OPEN: Self = _

The right to invoke path_open.

pub const FD_READDIR: Self = _

The right to invoke fd_readdir.

pub const PATH_READLINK: Self = _

The right to invoke path_readlink.

pub const PATH_RENAME_SOURCE: Self = _

The right to invoke path_rename with the file descriptor as the source directory.

pub const PATH_RENAME_TARGET: Self = _

The right to invoke path_rename with the file descriptor as the target directory.

pub const PATH_FILESTAT_GET: Self = _

The right to invoke path_filestat_get.

pub const PATH_FILESTAT_SET_SIZE: Self = _

The right to change a file’s size (there is no path_filestat_set_size). If rights::path_open is set, includes the right to invoke path_open with oflags::trunc.

pub const PATH_FILESTAT_SET_TIMES: Self = _

The right to invoke path_filestat_set_times.

pub const FD_FILESTAT_GET: Self = _

The right to invoke fd_filestat_get.

pub const FD_FILESTAT_SET_SIZE: Self = _

The right to invoke fd_filestat_set_size.

pub const FD_FILESTAT_SET_TIMES: Self = _

The right to invoke fd_filestat_set_times.

pub const PATH_SYMLINK: Self = _

The right to invoke path_symlink.

pub const PATH_REMOVE_DIRECTORY: Self = _

The right to invoke path_remove_directory.

pub const PATH_UNLINK_FILE: Self = _

The right to invoke path_unlink_file.

pub const POLL_FD_READWRITE: Self = _

If rights::fd_read is set, includes the right to invoke poll_oneoff to subscribe to eventtype::fd_read. If rights::fd_write is set, includes the right to invoke poll_oneoff to subscribe to eventtype::fd_write.

pub const SOCK_SHUTDOWN: Self = _

The right to invoke sock_shutdown.

pub const SOCK_ACCEPT: Self = _

TODO: Found in wasmer-wasi-types rust project, but not in wasi-snapshot0

pub const SOCK_CONNECT: Self = _

TODO: Found in wasmer-wasi-types rust project, but not in wasi-snapshot0

pub const SOCK_LISTEN: Self = _

TODO: Found in wasmer-wasi-types rust project, but not in wasi-snapshot0

pub const SOCK_BIND: Self = _

TODO: Found in wasmer-wasi-types rust project, but not in wasi-snapshot0

pub const SOCK_RECV: Self = _

TODO: Found in wasmer-wasi-types rust project, but not in wasi-snapshot0

pub const SOCK_SEND: Self = _

TODO: Found in wasmer-wasi-types rust project, but not in wasi-snapshot0

pub const SOCK_ADDR_LOCAL: Self = _

TODO: Found in wasmer-wasi-types rust project, but not in wasi-snapshot0

pub const SOCK_ADDR_REMOTE: Self = _

TODO: Found in wasmer-wasi-types rust project, but not in wasi-snapshot0

pub const SOCK_RECV_FROM: Self = _

TODO: Found in wasmer-wasi-types rust project, but not in wasi-snapshot0

pub const SOCK_SEND_TO: Self = _

TODO: Found in wasmer-wasi-types rust project, but not in wasi-snapshot0

pub const fn empty() -> Self

Returns an empty set of flags.

pub const fn all() -> Self

Returns the set containing all flags.

pub const fn bits(&self) -> u64

Returns the raw value of the flags currently stored.

pub const fn from_bits(bits: u64) -> Option<Self>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u64) -> Self

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const unsafe fn from_bits_unchecked(bits: u64) -> Self

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

Safety

The caller of the bitflags! macro can chose to allow or disallow extra bits for their bitflags type.

The caller of from_bits_unchecked() has to ensure that all bits correspond to a defined flag or that extra bits are valid for this bitflags type.

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: Self) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: Self) -> bool

Returns true if all of the flags in other are contained within self.

pub fn insert(&mut self, other: Self)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: Self)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: Self)

Toggles the specified flags in-place.

pub fn set(&mut self, other: Self, value: bool)

Inserts or removes the specified flags depending on the passed value.

pub const fn intersection(self, other: Self) -> Self

Returns the intersection between the flags in self and other.

Specifically, the returned set contains only the flags which are present in both self and other.

This is equivalent to using the & operator (e.g. ops::BitAnd), as in flags & other.

pub const fn union(self, other: Self) -> Self

Returns the union of between the flags in self and other.

Specifically, the returned set contains all flags which are present in either self or other, including any which are present in both (see Self::symmetric_difference if that is undesirable).

This is equivalent to using the | operator (e.g. ops::BitOr), as in flags | other.

pub const fn difference(self, other: Self) -> Self

Returns the difference between the flags in self and other.

Specifically, the returned set contains all flags present in self, except for the ones present in other.

It is also conceptually equivalent to the “bit-clear” operation: flags & !other (and this syntax is also supported).

This is equivalent to using the - operator (e.g. ops::Sub), as in flags - other.

pub const fn symmetric_difference(self, other: Self) -> Self

Returns the symmetric difference between the flags in self and other.

Specifically, the returned set contains the flags present which are present in self or other, but that are not present in both. Equivalently, it contains the flags present in exactly one of the sets self and other.

This is equivalent to using the ^ operator (e.g. ops::BitXor), as in flags ^ other.

pub const fn complement(self) -> Self

Returns the complement of this set of flags.

Specifically, the returned set contains all the flags which are not set in self, but which are allowed for this type.

Alternatively, it can be thought of as the set difference between Self::all() and self (e.g. Self::all() - self)

This is equivalent to using the ! operator (e.g. ops::Not), as in !flags.

impl Rights

pub fn from_bits_preserve(bits: u64) -> Self

Convert from a raw integer, preserving any unknown bits. See https://github.com/bitflags/bitflags/issues/263#issuecomment-957088321

impl Rights

pub const fn all_socket() -> Self

pub fn to_str(self) -> Option<&'static str>

expects a single right, returns None if out of bounds or > 1 bit set