wasmer_wasix_types::wasi::wasix_manual
Struct Event

#[repr(C)]
pub struct Event {
    pub userdata: Userdata,
    pub error: Errno,
    pub type_: Eventtype,
    pub u: EventUnion,
}

An event that occurred.

Fields
userdata: Userdata

User-provided value that got attached to subscription::userdata.

error: Errno

If non-zero, an error that occurred while processing the subscription request.

type_: Eventtype

Type of event that was triggered

u: EventUnion

The type of the event that occurred, and the contents of the event