wasmer_wasix_types::wasix

Struct WasiMemoryLayout

pub struct WasiMemoryLayout {
    pub stack_upper: u64,
    pub stack_lower: u64,
    pub guard_size: u64,
    pub stack_size: u64,
    pub tls_base: Option<u64>,
}

Represents the memory layout of the parts that the thread itself uses

Fields
stack_upper: u64

This is the top part of the stack (stacks go backwards)

stack_lower: u64

This is the bottom part of the stack (anything more below this is a stack overflow)

guard_size: u64

Piece of memory that is marked as none readable/writable so stack overflows cause an exception TODO: This field will need to be used to mark the guard memory as inaccessible

stack_size: u64

Total size of the stack

tls_base: Option<u64>

Pointer to the threadโs TLS (Thread Local Storage) memory, only filled in when itโs known. This is mainly used by the dynamic linker; especially so since we canโt even find the TLS area for a non-PIE module that doesnโt export its __tls_base. For DL modules, this is the TLS base of the main module.