Module wasmer_wasix_types::wasi::bindings

Structs

AddrUnspec

AddrUnspecPort

CidrUnspec

Dirent

A directory entry.

ErrnoSignal

Represents an errno and a signal

EventFdReadwrite

The contents of an event for the eventtype::fd_read and eventtype::fd_write variants

Eventrwflags

The state of the file descriptor subscribed to with eventtype::fd_read or eventtype::fd_write.

Fdflags

File descriptor flags.

Fdstat

File descriptor attributes.

Filestat

Fstflags

Which file time attributes to adjust. TODO: wit appears to not have support for flags repr (@witx repr u16)

HttpHandles

HttpStatus

JoinFlags

join flags.

Lookup

Flags determining the method of how paths are resolved. TODO: wit appears to not have support for flags repr (@witx repr u32)

Oflags

Open flags used by path_open. TODO: wit appears to not have support for flags repr (@witx repr u16)

OptionFd

OptionPid

OptionTimestamp

PipeHandles

Prestat

PrestatU

PrestatUDir

ProcessHandles

Rights

File descriptor rights, determining which actions may be performed.

Snapshot0Dirent

A directory entry, snapshot0 version.

Snapshot0Filestat

Snapshot0SubscriptionClock

The contents of a subscription when type is eventtype::clock.

Subclockflags

Flags determining how to interpret the timestamp provided in subscription-clock::timeout.

SubscriptionClock

The contents of a subscription when type is eventtype::clock.

SubscriptionFsReadwrite

An event that occurred. The contents of an event. An event that occurred. The contents of a subscription, snapshot0 version. The contents of a subscription. The contents of a subscription when the variant is eventtype::fd_read or eventtype::fd_write.

ThreadStateFlags

thread state flags

Tty

Enums

Addressfamily

Advice

File or memory access pattern advisory information.

Bool

Clockid

Identifiers for clocks.

Errno

Error codes returned by functions. Not all of these error codes are returned by the functions provided by this API; some are used in higher-level library layers, and others are provided merely for alignment with POSIX.

Eventtype

Type of a subscription to an event or its occurrence.

Filetype

The type of a file descriptor or file.

JoinStatusType

What has happened with the proccess when we joined on it

OptionTag

__wasi_option_t

Preopentype

Identifiers for preopened capabilities.

Signal

Snapshot0Clockid

Identifiers for clocks, snapshot0 version.

Snapshot0Whence

SockProto

Sockoption

Sockstatus

Socktype

StdioMode

Streamsecurity

Timeout

Whence

Constants

Monotonic__num_enum_0__ ğŸ”’

Monotonic__num_enum_0__ ğŸ”’

ProcessCputimeId__num_enum_0__ ğŸ”’

ProcessCputimeId__num_enum_0__ ğŸ”’

Realtime__num_enum_0__ ğŸ”’

Realtime__num_enum_0__ ğŸ”’

Sigabrt__num_enum_0__ ğŸ”’

Sigalrm__num_enum_0__ ğŸ”’

Sigbus__num_enum_0__ ğŸ”’

Sigchld__num_enum_0__ ğŸ”’

Sigcont__num_enum_0__ ğŸ”’

Sigfpe__num_enum_0__ ğŸ”’

Sighup__num_enum_0__ ğŸ”’

Sigill__num_enum_0__ ğŸ”’

Sigint__num_enum_0__ ğŸ”’

Sigkill__num_enum_0__ ğŸ”’

Signone__num_enum_0__ ğŸ”’

Sigpipe__num_enum_0__ ğŸ”’

Sigpoll__num_enum_0__ ğŸ”’

Sigprof__num_enum_0__ ğŸ”’

Sigpwr__num_enum_0__ ğŸ”’

Sigquit__num_enum_0__ ğŸ”’

Sigsegv__num_enum_0__ ğŸ”’

Sigstkflt__num_enum_0__ ğŸ”’

Sigstop__num_enum_0__ ğŸ”’

Sigsys__num_enum_0__ ğŸ”’

Sigterm__num_enum_0__ ğŸ”’

Sigtrap__num_enum_0__ ğŸ”’

Sigtstp__num_enum_0__ ğŸ”’

Sigttin__num_enum_0__ ğŸ”’

Sigttou__num_enum_0__ ğŸ”’

Sigurg__num_enum_0__ ğŸ”’

Sigusr1__num_enum_0__ ğŸ”’

Sigusr2__num_enum_0__ ğŸ”’

Sigvtalrm__num_enum_0__ ğŸ”’

Sigwinch__num_enum_0__ ğŸ”’

Sigxcpu__num_enum_0__ ğŸ”’

Sigxfsz__num_enum_0__ ğŸ”’

ThreadCputimeId__num_enum_0__ ğŸ”’

ThreadCputimeId__num_enum_0__ ğŸ”’

Unknown__num_enum_0__ ğŸ”’

Unknown__num_enum_0__ ğŸ”’

Statics

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

__CALLSITE ğŸ”’

Traits

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

__BitFlags ğŸ”’

Type Definitions

Count

Device

Identifier for a device containing a file system. Can be used in combination with inode to uniquely identify a file or directory in the filesystem.

Dircookie

A reference to the offset of a directory entry.

Dirnamlen

The type for the dirent::d-namlen field of dirent struct.

EventFdFlags

Fd

A file descriptor handle.

FileDelta

Filesize

Non-negative file size or length of a region within a file.

Inode

File serial number that is unique within its file system.

Linkcount

LookupFlags

Pid

RiFlags

RoFlags

SdFlags

SiFlags

Size

Type names used by low-level WASI interfaces. An array size.

Snapshot0Linkcount

Tid

Timestamp

Timestamp in nanoseconds.

Userdata

User-provided value that may be attached to objects that is retained when extracted from the implementation.